Set short format version as default

+++ Covid-19 Status/src/Preferences/PreferencesOptions.swift
@@ -47,5 +47,5 @@ class PreferencesOptions {
 
     static var defaultAlertsStatus = true
 
-    static var defaultFormatMethod = StatsFormatMethod.long
+    static var defaultFormatMethod = StatsFormatMethod.short
 }
